Commit 5
Added support for customers to select a time and date for order. I used helpers that are browser specific unfortunately. If a browser is unsupported, it formats to a normal text field. They do not work on Firefox on my machine, not sure why but they do work on Safari on my phone and will most likely work on Android as well.

TODO:

Add support for images on meals.
Add support for discount on orders.
Add validation for time and date orders - times before the current can be chosen by the user.